Relocating team-view components
