+++ src/app/header/header.component.spec.ts
@@ -11,7 +11,7 @@ import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {ConfigurationService} from '../config/configuration.service';
 import {Configuration} from '../config/configuration.model';
 import {first} from 'rxjs/operators';
-import {TeamViewDashboardService} from '../twitch/twitch-dashboard/team-view-dashboard.service';
+import {TeamViewDashboardService} from '../team-view/team-view-dashboard.service';
 import {MockTwitchDashboardService} from '../test/mocks/twitch-service.mock.spec';
 import {BreakpointObserver, BreakpointState} from '@angular/cdk/layout';
 import {Observable, of} from 'rxjs';
