+++ src/app/team-view/team-view-dashboard.component.scss
