+++ src/app/team-view/twitch-card-measurements.model.ts
