+++ src/app/test/mocks/twitch-service.mock.spec.ts
@@ -1,6 +1,6 @@
 import {of} from 'rxjs';
 import createSpyObj = jasmine.createSpyObj;
-import {TeamViewDashboardService} from '../../twitch/twitch-dashboard/team-view-dashboard.service';
+import {TeamViewDashboardService} from '../../team-view/team-view-dashboard.service';
 import {TournamentDashboardTwitchService} from '../../tournament/tournament-dashboard/tournament-dashboard-twitch.service';
 
 // tslint:disable-next-line:variable-name
